requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
PyJWT==2.1.0
python-decouple==3.5
python-slugify==5.0.2
pytz==2021.3
sqlparse==0.4.2
text-unidecode==1.3
